encode/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
cffi==1.16.0
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
ffmpeg-python==0.2.0
flatbuffers==23.5.26
fonttools==4.47.2
future==0.18.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mediapipe==0.10.9
nest-asyncio==1.6.0
numpy==1.24.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
vulture==2.11
wcwidth==0.2.13
XlsxWriter==3.2.0
zipp==3.17.0
